Update AbstractGateway.php
This PR addresses deprecation warnings in PHP 8.2 regarding the creation of dynamic properties in the AbstractGateway class used in the btcpay-greenfield-for-woocommerce plugin.
Problem:

PHP 8.2 introduced deprecation warnings for the creation of dynamic properties in objects that do not explicitly define them. In the AbstractGateway class, the properties $debug_php_version and $debug_plugin_version were being dynamically added in the constructor, which causes the following warning:

Creation of dynamic property BTCPayServer\WC\Gateway\DefaultGateway::$debug_plugin_version is deprecated.

Added explicit class properties for $debug_php_version and $debug_plugin_version to prevent PHP 8.2 from issuing deprecation warnings.
The properties are declared as protected and their values are initialized in the constructor, as they were previously.